[EC-28] refactor: LeftLineItem 인라인 스타일 수정
