+++ client/src/components/listItem/leftLineListItem/LeftLineListItem.jsx
@@ -1,24 +1,17 @@
 import React from 'react'
 import styles from "./LeftLineListItem.module.css"
-import { getBackgroundColor } from '../../../utils/absenceAttendancesList';
+import { getBackgroundColor } from '../../../utils/buttonContentList';
 import MoreButton from '../../buttons/moreButton/MoreButton';
 
+// TODO : 관리자의 유고 결석 관리만 isClickable = true
 export default function LeftLineListItem({ isClickable, handleClick, status, children }) {
-  const backgroundColor = getBackgroundColor(status);
+  const bgColor = getBackgroundColor(status);
 
-  return isClickable ? (
-    <button
-      disabled={!isClickable}
-      onClick={handleClick}
-      style={{ '--after-bg': backgroundColor }}
-      className={styles.leftLineListItem}
+  return (
+    <div
+      className={`${styles.leftLineListItem} ${isClickable && `${styles.active}`} ${bgColor ? styles[bgColor] : ''}`}
     >
       {children}
-    </button>
-  ) : (
-    <div style={{ '--after-bg': backgroundColor }} className={styles.leftLineListItem}>
-      {children}
-      <div>wldnf</div>
 
       {!isClickable && <MoreButton></MoreButton>}
     </div>
